% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_rules.R
\name{info_rules}
\alias{info_rules}
\title{Displaya information about file extension names for GNU Make pattern rules}
\usage{
info_rules(dependency.ext = "R", list.all = FALSE)
}
\arguments{
\item{dependency.ext}{\code{string} specifying the filename
extension to obtain a description of possible target file
extensions. Default: \dQuote{R}}

\item{list.all}{\code{logical} indicating whether to list all
possible dependency filename extensions. Default: FALSE}
}
\value{
Nothing is returned but instead information is printed
}
\description{
\code{info_rules} displays target filename extensions for dependency files
}
\details{
GNU Make pattern rules allow wildcarding so that generic rules may
be reused based on the filename extensions of the target file and
first dependecy file. \code{info_rules} aids construction of
Makefiles by displaying possible target file name extensions given
a dependency file name extension. Alternatively, it can list all
possible dependency file name extensions employed with the
\code{\link{gnumaker}} package.
}
\examples{
info_rules(list.all = TRUE)
info_rules("Rmd")

}
\seealso{
\code{\link{create_makefile}}
}
