% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gnu_makefile.R
\name{plot.gnu_makefile}
\alias{plot.gnu_makefile}
\title{Plots DAG from a GNU Makefile object}
\usage{
\method{plot}{gnu_makefile}(
  x,
  nodes.color = "wheat",
  parents.shape = "box",
  parents.color = "lightgreen",
  ...
)
}
\arguments{
\item{x}{object of class \code{gnu_makefile}}

\item{nodes.color}{color of intermediate files. Default \dQuote{wheat}}

\item{parents.shape}{shape of parent files box. Default \dQuote{box}}

\item{parents.color}{color of parent files. Default \dQuote{lightgreen}}

\item{\\dots}{options passed to  \code{Rgraphviz} \code{\link[Rgraphviz]{plot.graphNEL}}}
}
\value{
does not return an object but is used for plotting
}
\description{
Standard GNU Make pattern rules are available at
https://github.com/petebaker/r-makefile-definitions and also with
the \code{gnumaker} package. Plots directed acyclic graph (DAG)
created using \code{\link{create_makefile}} using package
\code{Rgraphviz}.
}
\examples{
# Create a Makfile for simple data analysis demo
gm1 <-
  create_makefile(targets = list(read = c("read.R", "simple.csv"),
                  linmod = c("linmod.R", "read"),
                  rep1 = c("report1.Rmd", "linmod"),
                  rep2 = c("report2.Rmd", "linmod")),
                  target.all = c("rep1", "rep2"),
                  all.exts = list(rep1 = "pdf", rep2 = "docx"),
                  comments = list(linmod = "plots and analysis using 'linmod.R'"))
plot(gm1)
plot(gm1, main = "Makefile for Simple Demo")
plot(gm1, attrs = list(node = list(fillsize = "20",
     edge = list(weight = 2.0))))

}
\seealso{
\code{\link{create_makefile}} \code{\link[Rgraphviz]{plot.graphNEL}}
}
