% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_makefile.R
\name{print.gnu_makefile}
\alias{print.gnu_makefile}
\title{Print an object of class gnu_makefile}
\usage{
\method{print}{gnu_makefile}(x, ...)
}
\arguments{
\item{x}{of class gnu_makefile}

\item{...}{generic optional arguments to print function}
}
\description{
Print an object of class gnu_makefile
}
